0.3.3

+++ pages/5_show_rented_books.py
@@ -20,12 +20,12 @@ def show_rented_books():
         st.title("Mis Libros Prestados")
         for book in rented_books:
             with st.container(border=True):
-                st.write(f"**Título:** {book['book_title']}")
-                if st.button(f"Devolver {book['book_title']}", key=f"return_{book['book_title']}"):
+                st.write(f"**Título:** {book['title']}")
+                if st.button(f"Devolver {book['title']}", key=f"return_{book['title']}"):
                     # Llamar directamente a la función de devolución sin usar el diálogo
-                    success = return_book(book['book_title'], st.session_state.user["id"])
+                    success = return_book(book['title'], st.session_state.user["id"])
                     if success:
-                        st.session_state.notification_message = f"Libro '{book['book_title']}' devuelto correctamente."
+                        st.session_state.notification_message = f"Libro '{book['title']}' devuelto correctamente."
                         st.session_state.show_notification = True
 
                         st.rerun()  # Usando st.rerun() en lugar de experimental_rerun
@@ -41,15 +41,12 @@ def return_book(book_title, user_id):
     try:
         conn = st.connection('biblionline_db', type='sql')
         with conn.session as s:
-            # Consulta SQL para eliminar el libro prestado
-            query = text("""
-                DELETE FROM rented_books 
-                WHERE user_id = :user_id AND book_title = :book_title
-            """)
-            s.execute(query, {
-                'user_id': user_id,
-                'book_title': book_title
-            })
+            
+            s.execute("DELETE FROM rented_books WHERE user_id = :user_id AND book_title = :book_title", 
+                      {
+                        'user_id': user_id,
+                        'book_title': book_title
+                        })
             s.commit()
         return True
     except Exception as e:
